requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
sqladmin>=0.11.0
sqlalchemy>=2.0.0
alembic>=1.10.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0
requests>=2.28.0
